testdrive: split kafka actions into modules
The main Kafka module was getting rather unwieldy.
